chore(data-modeling): relocate api views into presentation/views
First step of data_modeling isolation: moves backend/api/{dag,edge,node}.py into backend/presentation/views/ (and api/test into tests/api), rewriting dotted paths repo-wide via hogli product:isolate:move. Views keep import-linter allowances to reach internals for now; they route through the facade in the presentation-wave follow-up.
