--- products/data_modeling/backend/api/__init__.py
@@ -1,5 +1,0 @@
-from .dag import DAGViewSet
-from .edge import EdgeViewSet
-from .node import NodeViewSet
-
-__all__ = ["DAGViewSet", "EdgeViewSet", "NodeViewSet"]
