+++ products/data_modeling/backend/presentation/__init__.py
@@ -1,0 +1,1 @@
+"""HTTP presentation layer of the data_modeling product (DRF viewsets and serializers)."""
