+++ products/data_modeling/backend/presentation/views/__init__.py
@@ -1,0 +1,5 @@
+from products.data_modeling.backend.presentation.views.dag import DAGViewSet
+from products.data_modeling.backend.presentation.views.edge import EdgeViewSet
+from products.data_modeling.backend.presentation.views.node import NodeViewSet
+
+__all__ = ["DAGViewSet", "EdgeViewSet", "NodeViewSet"]
